+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.20.33
+
+* `Colors`
+    * `Common`:
+        * Add opportunity to use `HEXAColor` with `ahex` colors
+
 ## 0.20.32
 
 * `Versions`: